requirements.txt
pyserial
qasync